Segregated creation of Issue & Fork from enricher into separate action
