+++ tools/check-disclosures/assets/schemas/vulnerability.json
@@ -1,15 +1,27 @@
 {
   "type": "object",
   "properties": {
-    "PackageVulnerabilityID": { "type": "string" },
-    "DisclosureDate": { "type": "string" },
-    "AffectedVersionRange": { "type": "string" },
-    "Summary": { "type": "string" },
+    "PackageVulnerabilityID": {
+      "type": "string"
+    },
+    "DisclosureDate": {
+      "type": "string"
+    },
+    "AffectedVersionRange": {
+      "type": "string"
+    },
+    "Summary": {
+      "type": "string"
+    },
     "Author": {
       "type": "object",
       "properties": {
-        "Username": { "type": "string" },
-        "Name": { "type": "string" }
+        "Username": {
+          "type": "string"
+        },
+        "Name": {
+          "type": "string"
+        }
       },
       "required": [
         "Username"
@@ -18,10 +30,18 @@
     "Package": {
       "type": "object",
       "properties": {
-        "Registry": { "type": "string" },
-        "Name": { "type": "string" },
-        "URL": { "type": "string" },
-        "Downloads": { "type": "string" }
+        "Registry": {
+          "type": "string"
+        },
+        "Name": {
+          "type": "string"
+        },
+        "URL": {
+          "type": "string"
+        },
+        "Downloads": {
+          "type": "string"
+        }
       },
       "required": [
         "Registry",
@@ -31,32 +51,60 @@
     },
     "CWEs": {
       "type": "array",
-      "items": [
-        {
-          "type": "object",
-          "properties": {
-            "ID": { "type": "string" },
-            "Description": { "type": "string" }
+      "items": [{
+        "type": "object",
+        "properties": {
+          "ID": {
+            "type": "string"
+          },
+          "Description": {
+            "type": "string"
           }
         }
-      ]
+      }]
     },
     "CVSS": {
       "type": "object",
       "properties": {
-        "Version": { "type": "string" },
-        "AV": { "type": "string" },
-        "AC": { "type": "string" },
-        "PR": { "type": "string" },
-        "UI": { "type": "string" },
-        "S": { "type": "string" },
-        "C": { "type": "string" },
-        "I": { "type": "string" },
-        "A": { "type": "string" },
-        "E": { "type": "string" },
-        "RL": { "type": "string" },
-        "RC": { "type": "string" },
-        "Score": { "type": "string" }
+        "Version": {
+          "type": "string"
+        },
+        "AV": {
+          "type": "string"
+        },
+        "AC": {
+          "type": "string"
+        },
+        "PR": {
+          "type": "string"
+        },
+        "UI": {
+          "type": "string"
+        },
+        "S": {
+          "type": "string"
+        },
+        "C": {
+          "type": "string"
+        },
+        "I": {
+          "type": "string"
+        },
+        "A": {
+          "type": "string"
+        },
+        "E": {
+          "type": "string"
+        },
+        "RL": {
+          "type": "string"
+        },
+        "RC": {
+          "type": "string"
+        },
+        "Score": {
+          "type": "string"
+        }
       },
       "required": [
         "Version",
@@ -69,7 +117,15 @@
     "Repository": {
       "type": "object",
       "properties": {
-        "URL": { "type": "string" },
+        "URL": {
+          "type": "string"
+        },
+        "Owner": {
+          "type": "string"
+        },
+        "Name": {
+          "type": "string"
+        },
         "Codebase": {
           "type": "array"
         }
@@ -87,7 +143,7 @@
     }
   },
   "required": [
-    "PackageVulnerabilityID", 
+    "PackageVulnerabilityID",
     "DisclosureDate",
     "AffectedVersionRange",
     "Summary",
